refactor(parser): split identifier metadata manager
